Set "reasonable" limits for brace nesting
The essential problem is: every time you write `$$x$}`, you get another
entry added to a hash table. Even if it's not [theoretically] *quadratic*,
it's still slow. Hard limiting it to 255 entries makes this not a problem.

Interestingly enough, when I tried to write an analogous torture test
for code spans, I couldn't find a way to do it because code spans are
keyed by their *length* instead of their *position*. In order to get
N entries in the hash table, I basically had to write N `` ` `` in a
row, forcing me to write quadratic amounts of input text.

Comparison:

```
michaelhowell@Michael-Howells-Macbook-Pro pulldown-cmark % python3 -c 'print("$$x$}"*5000)' | time target/release/pulldown-cmark.old -M > /dev/null
target/release/pulldown-cmark.old -M > /dev/null  2.63s user 0.02s system 99% cpu 2.673 total
michaelhowell@Michael-Howells-Macbook-Pro pulldown-cmark % python3 -c 'print("$$x$}"*5000)' | time target/release/pulldown-cmark.new -M > /dev/null
target/release/pulldown-cmark.new -M > /dev/null  0.01s user 0.00s system 6% cpu 0.109 total
```

[theoretically]: http://www.ilikebigbits.com/2014_04_21_myth_of_ram_1.html

+++ pulldown-cmark/src/parse.rs
@@ -63,7 +63,7 @@ pub(crate) enum ItemBody {
     // repeats, can_open, can_close
     MaybeEmphasis(usize, bool, bool),
     // is_display, can_open, can_close, brace context
-    MaybeMath(bool, bool, usize),
+    MaybeMath(bool, bool, u8),
     // quote byte, can_open, can_close
     MaybeSmartQuote(u8, bool, bool),
     MaybeCode(usize, bool), // number of backticks, preceded by backslash
@@ -1626,7 +1626,7 @@ impl CodeDelims {
 /// Tracks brace contexts and delimiter length for math delimiters.
 /// Provides amortized constant-time lookups.
 struct MathDelims {
-    inner: HashMap<(usize, usize), VecDeque<(TreeIndex, bool, bool)>>,
+    inner: HashMap<(usize, u8), VecDeque<(TreeIndex, bool, bool)>>,
     seen_first: bool,
 }
 
@@ -1638,7 +1638,7 @@ impl MathDelims {
         }
     }
 
-    fn insert(&mut self, delim_count: usize, brace_context: usize, ix: TreeIndex, can_close: bool, delim_is_display: bool) {
+    fn insert(&mut self, delim_count: usize, brace_context: u8, ix: TreeIndex, can_close: bool, delim_is_display: bool) {
         if self.seen_first {
             self.inner
                 .entry((delim_count, brace_context))
@@ -1655,7 +1655,7 @@ impl MathDelims {
         !self.inner.is_empty()
     }
 
-    fn find(&mut self, open_ix: TreeIndex, mut delim_count: usize, brace_context: usize, is_display: bool) -> Option<(TreeIndex, bool)> {
+    fn find(&mut self, open_ix: TreeIndex, mut delim_count: usize, brace_context: u8, is_display: bool) -> Option<(TreeIndex, bool)> {
         while delim_count > 0 {
             while let Some((ix, can_close, delim_is_display)) = self.inner.get_mut(&(delim_count, brace_context))?.pop_front() {
                 if ix > open_ix {
